pyproject.toml
[project]
name = "finance-stock-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.0",
    "google-adk>=1.18.0",
    "instructor>=1.13.0",
    "openai>=2.8.1",
    "peewee>=3.18.3",
    "playhouse>=0.1.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
